# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "camp-backend",
    "description": "\n    # AIエージェントアプリケーション API\n\n    LangChain/LangGraphを使用したAIエージェントアプリケーションのバックエンドAPI。\n\n    ## 認証モード: development\n\n    ### 開発モード認証（モック）\n\n    ## 主な機能\n\n    - **AIエージェントとのチャット**: LangGraphによる高度な会話エンジン\n    - **セッション管理**: 会話履歴の保存と取得\n    - **ファイル処理**: ファイルのアップロード・ダウンロード\n    - **認証・認可**: JWT認証 とロールベース制御\n    - **モニタリング**: Prometheusメトリクス収集\n\n    ## 認証\n\n    ほとんどのエンドポイントはゲストアクセス可能ですが、一部の機能では認証が必要です。\n    認証が必要な場合は、`Authorization: Bearer <token>` ヘッダーを含めてください。\n\n    ## レート制限\n\n    APIリクエストは100リクエスト/分に制限されています。\n    ",
    "contact": {
      "name": "開発チーム",
      "email": "nomura.koshiro@gmail.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.2"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト一覧取得",
        "description": "認証されたユーザーが所属するプロジェクトの一覧を取得します。\n\n    **認証が必要です。**\n\n    - 自分がメンバーとして所属するプロジェクトのみ表示されます\n    - ロール（OWNER/ADMIN/MEMBER/VIEWER）に関係なく全て取得されます\n\n    クエリパラメータ:\n        - skip: スキップするレコード数（デフォルト: 0）\n        - limit: 取得する最大レコード数（デフォルト: 100、最大: 1000）\n        - is_active: アクティブフラグフィルタ（オプション）",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "スキップするレコード数",
              "default": 0,
              "title": "Skip"
            },
            "description": "スキップするレコード数"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "取得する最大レコード数",
              "default": 100,
              "title": "Limit"
            },
            "description": "取得する最大レコード数"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "アクティブフラグフィルタ",
              "title": "Is Active"
            },
            "description": "アクティブフラグフィルタ"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Api V1 Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト作成",
        "description": "新しいプロジェクトを作成します。\n\n    **認証が必要です。**\n\n    - 作成者は自動的にOWNERロールとしてプロジェクトメンバーに追加されます\n    - プロジェクトコードは一意である必要があります\n\n    リクエストボディ:\n        - name: プロジェクト名（最大255文字）\n        - code: プロジェクトコード（最大50文字、一意）\n        - description: プロジェクト説明（オプション）",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト詳細取得",
        "description": "指定されたIDのプロジェクト情報を取得します。\n\n    **認証が必要です。**\n    **プロジェクトメンバーのみアクセス可能です。**\n\n    パスパラメータ:\n        - project_id: プロジェクトID（UUID形式）",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト情報更新",
        "description": "プロジェクト情報を更新します。\n\n    **OWNER/ADMINロールが必要です。**\n\n    更新可能なフィールド:\n        - name: プロジェクト名\n        - description: プロジェクト説明\n        - is_active: アクティブフラグ\n\n    注意:\n        - code（プロジェクトコード）は更新できません\n        - 指定されたフィールドのみが更新されます（PATCH）",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト削除",
        "description": "プロジェクトを削除します。\n\n    **OWNERロールが必要です。**\n\n    パスパラメータ:\n        - project_id: プロジェクトID（UUID形式）\n\n    注意:\n        - 削除は物理削除です（データベースから完全に削除されます）\n        - CASCADE設定により、関連する ProjectMember と ProjectFile も自動削除されます\n        - この操作は取り消せません",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/members": {
      "get": {
        "tags": [
          "project-members"
        ],
        "summary": "プロジェクトメンバー一覧取得",
        "description": "プロジェクトのメンバー一覧を取得します。\n\n    **メンバー以上の権限が必要です。**\n\n    - ユーザー情報付きで返されます\n    - ページネーション対応\n\n    クエリパラメータ:\n        - skip: スキップするレコード数（デフォルト: 0）\n        - limit: 取得する最大レコード数（デフォルト: 100、最大: 1000）",
        "operationId": "get_project_members_api_v1_projects__project_id__members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "project-members"
        ],
        "summary": "プロジェクトメンバー追加",
        "description": "プロジェクトに新しいメンバーを追加します。\n\n    **PROJECT_MANAGERの権限が必要です。**\n\n    - 重複するメンバーは追加できません\n\n    リクエストボディ:\n        - user_id: 追加するユーザーのUUID\n        - role: プロジェクトロール（project_manager/member/viewer）",
        "operationId": "add_project_member_api_v1_projects__project_id__members_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectMemberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/members/me": {
      "get": {
        "tags": [
          "project-members"
        ],
        "summary": "自分のロール取得",
        "description": "プロジェクトにおける自分のロールを取得します。\n\n    **メンバー以上の権限が必要です。**\n\n    返されるデータ:\n        - project_id: プロジェクトUUID\n        - user_id: ユーザーUUID\n        - role: プロジェクトロール\n        - is_owner: PROJECT_MANAGER ロールかどうか（後方互換性のため維持）\n        - is_admin: PROJECT_MANAGER ロールかどうか（後方互換性のため維持）",
        "operationId": "get_my_role_api_v1_projects__project_id__members_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "project-members"
        ],
        "summary": "プロジェクト退出",
        "description": "プロジェクトから自分自身を退出します。\n\n    **任意のメンバーが実行可能です。**\n\n    - 最後の PROJECT_MANAGER は退出できません",
        "operationId": "leave_project_api_v1_projects__project_id__members_me_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectMemberCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "追加するユーザーのID"
          },
          "role": {
            "$ref": "#/components/schemas/ProjectRole",
            "description": "プロジェクトロール（project_manager/project_moderator/member/viewer）",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ProjectMemberCreate",
        "description": "プロジェクトメンバー追加リクエストスキーマ。\n\nプロジェクトに新しいメンバーを追加する際に使用します。\n\nAttributes:\n    user_id (uuid.UUID): 追加するユーザーのID\n    role (ProjectRole): プロジェクトロール（デフォルト: MEMBER）\n\nExample:\n    >>> member = ProjectMemberCreate(\n    ...     user_id=uuid.uuid4(),\n    ...     role=ProjectRole.MEMBER\n    ... )\n\nNote:\n    - PROJECT_MANAGER 権限が必要\n    - 重複するメンバーは追加できません"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}